% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JNNI.R
\name{GANAIr}
\alias{GANAIr}
\title{fit a Bayesian model for the integration of nodes, interaction terms and networks}
\usage{
GANAIr(X, pathway, L = NULL, t, s2, prob)
}
\arguments{
\item{X}{a matrix of predictor variable}

\item{pathway}{a vector of pathway information, for instance, value 1 represent the corresponding node location belongs to pathway 1}

\item{L}{a list of Laplace information, each of which is a matrix capturing the relationship between node in the corresponding pathway, default is NULL}

\item{t}{a vector of the response variable}

\item{s2}{a numeric}

\item{prob}{a numeric}
}
\value{
A list containing values as follows:
\item{mean_w_red}{The coefficients vector}
\item{beta_interaction}{The coefficients of interaction terms}
\item{beta_main}{The coefficients of main effects}
\item{BIC}{Bayesian Information Criterion}
\item{df}{The number of nonzero coefficients}
\item{index_path}{Selected networks' number}
}
\description{
One of the main functions in the JNNI package.
}
\examples{
library(JNNI)
set.seed(1)
X=matrix(rnorm(50),10,5)
pathway=c(1,1,2,2,3)
b=c(3,3,0,0,0)
t=as.numeric(X\%*\%b+rnorm(10))
s2=1e-3
prob=0.96
temp=GANAIr(X,pathway,L=NULL,t,s2,prob)
}
